refactor : DB로 전환 - list - ask - show  메서드 변경 -> 기존 List에서 Repository로의 변환에 따라
show.html -> modelAttribute 값 템플릿 엔진 변수이름 수정
