+++ src/main/java/codesquad/domain/question/Question.java
@@ -1,10 +1,25 @@
-package codesquad.domain;
+package codesquad.domain.question;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
 public class Question {
-	private String writer;
-	private String title;
-	private String contents;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private long index;
+
+	@Column(nullable = false, length = 20)
+	private String writer;
+
+	@Column(nullable = false, length = 200)
+	private String title;
+
+	@Column(nullable = false, columnDefinition = "TEXT")
+	private String contents;
 
 	public Long getIndex() {
 		return index;
