+++ src/main/java/codesquad/domain/question/QuestionRepository.java
@@ -1,0 +1,7 @@
+package codesquad.domain.question;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface QuestionRepository extends CrudRepository<Question, Long> {
+
+}
