+++ src/main/resources/templates/qna/show.html
@@ -1,7 +1,7 @@
 {{>layout/header}}
 
 <div class="container" id="main">
-    {{#questions}}
+    {{#question}}
     <div class="col-md-12 col-sm-12 col-lg-12">
         <div class="panel panel-default">
           <header class="qna-header">
